Changed the way Actionmodules add users to Group/Course
The function to add some user to a group or a course is not anymore in
the modules itself. This was very redundant. So I created some central
functions for this in the ActionModuleBase class.

From now on, the person who executes the import will not be added to the
courses/groups. If he/she wishes to be added, he/she have to add
him-/herself to the list of admins who will be added.